backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-multipart
pypdf
pandas
regex
requests



passlib[argon2]
python-jose[cryptography]
